pyproject.toml
[project]
name = "VideoQuery"
version = "0.1.0"
description = "A tool that uses an LLM to search through a video's audio content."
authors = [
    {name = "jimdel",email = "jdelay.jr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "yt-dlp (>=2023.12.30)",
    "openai-whisper (>=20231117)",
    "langchain (>=0.1.0)",
    "langchain-community (>=0.0.10)",
    "langchain-openai (>=0.0.2)",
    "chromadb (>=0.4.22)",
    "openai (>=1.12.0)",
    "tiktoken (>=0.5.2)",
    "python-dotenv (>=1.0.0)"
]

[tool.poetry]
packages = [{include = "video_query", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
